card_image_generator/requirements.txt
flask
pyyaml
pillow
edn_format
waitress
